Stop PRs to same repo interfering with each other
Previously the branch name was static,
which caused issues when terraform opens PRs to multiple target branches
to the same repository.
This is because the terraform run for each target branch
was trying to work on the same branch,
and opening PRs from the same branch.
This was observed in repos when we maintain multiple branches
(eg. snap-tempest with main, stable/caracal, stable/bobcat, etc.).

This fixes it by including the target branch name in the PR branch,
ensuring that each PR in a target repository will use its own branch,
and there won't be any more cross interference.

Fixes: #88