Add charmcraft and python versions to charm check
Some projects now use charmcraft 3, while many still use charmcraft 2,
so we need a variable for the charm check.yaml template now.

Similarly for python versions - it was hardcoded to 3.8 and 3.10,
but some projects don't support python 3.8 (focal) any more,
so we only run tests on python 3.10 (jammy).
So we need a variable there to manage python versions per-project.